+++ tests/testthat/test-wflow_open.R
@@ -1,0 +1,116 @@
+context("wflow_open")
+
+# Setup ------------------------------------------------------------------------
+
+# start project in a tempdir
+site_dir <- tempfile("test-wflow_open-")
+suppressMessages(wflow_start(site_dir, change_wd = FALSE))
+site_dir <- workflowr:::relative(site_dir)
+on.exit(unlink(site_dir, recursive = TRUE, force = TRUE))
+p <- workflowr:::wflow_paths(project = site_dir)
+
+# Test wflow_open --------------------------------------------------------------
+
+test_that("wflow_open creates a new file, but does not overwrite", {
+
+  rmd <- wflow_open(file.path(p$analysis, "test.Rmd"), change_wd = FALSE,
+                    open_file = FALSE, project = site_dir)
+  expect_true(file.exists(rmd))
+  modification_time_pre <- file.mtime(rmd)
+  Sys.sleep(2)
+  rmd2 <- wflow_open(file.path(p$analysis, "test.Rmd"), change_wd = FALSE,
+                     open_file = FALSE, project = site_dir)
+  expect_identical(rmd2, rmd)
+  modification_time_post <- file.mtime(rmd2)
+  expect_identical(modification_time_post, modification_time_pre)
+})
+
+test_that("wflow_open changes the working directory", {
+
+  wd_pre <- getwd()
+  on.exit(setwd(wd_pre))
+  expect_silent(rmd <- wflow_open(file.path(p$analysis, "do-not-change-wd.Rmd"), change_wd = FALSE,
+                                  open_file = FALSE, project = site_dir))
+  expect_true(file.exists(rmd))
+  wd_same <- getwd()
+  expect_identical(wd_pre, wd_same)
+  expect_message(rmd <- wflow_open(file.path(p$analysis, "do-change-wd.Rmd"), change_wd = TRUE,
+                                   open_file = FALSE, project = site_dir),
+                 "Working directory was changed.")
+  expect_true(file.exists(rmd))
+  wd_change <- workflowr:::relative(getwd(), start = wd_pre)
+  expect_identical(wd_change, file.path(site_dir, "analysis"))
+  expect_silent(rmd <- wflow_open("no-need-to-change-wd.Rmd",
+                                  change_wd = TRUE, open_file = FALSE,
+                                  project = "."))
+  expect_true(file.exists(rmd))
+})
+
+test_that("wflow_open can accept multiple files", {
+
+  rmd_multi <- file.path(p$analysis, paste0(1:3, ".Rmd"))
+  rmd <- wflow_open(rmd_multi, change_wd = FALSE, open_file = FALSE,
+                    project = site_dir)
+  expect_true(all(file.exists(rmd)))
+  modification_time_pre <- file.mtime(rmd)
+  Sys.sleep(2)
+  rmd2 <- wflow_open(rmd_multi, change_wd = FALSE, open_file = FALSE,
+                     project = site_dir)
+  expect_identical(rmd2, rmd)
+  modification_time_post <- file.mtime(rmd)
+  expect_identical(modification_time_post, modification_time_pre)
+})
+
+# test_that("wflow_open can accept basename, full paths, and wrong paths", {
+#
+#   rmd_paths <- c("basename.Rmd",
+#                  file.path(site_dir, "analysis", "full.Rmd"),
+#                  file.path(site_dir, "code", "wrong.Rmd"))
+#   rmd <- wflow_open(rmd_paths, change_wd = FALSE, open_file = FALSE,
+#                     project = site_dir)
+#   expect_true(all(file.exists(rmd)))
+#   modification_time_pre <- file.mtime(rmd)
+#   Sys.sleep(2)
+#   rmd2 <- wflow_open(rmd_paths, change_wd = FALSE, open_file = FALSE,
+#                      project = site_dir)
+#   expect_identical(rmd2, rmd)
+#   modification_time_post <- file.mtime(rmd)
+#   expect_identical(modification_time_post, modification_time_pre)
+# })
+
+test_that("wflow_open can save outside of analysis/ when project = NULL", {
+
+  # When project = NULL, wflow_open will create output directories if needed, and
+  # switches the working directory to the path of the first input file.
+  cwd <- getwd()
+  on.exit(setwd(cwd))
+  location_exist <- tempfile("test-wflow_open-exist-")
+  dir.create(location_exist)
+  location_exist <- workflowr:::absolute(location_exist)
+  on.exit(unlink(location_exist, recursive = TRUE, force = TRUE), add = TRUE)
+  testfile1 <- file.path(location_exist, "exist.Rmd")
+  location_nonexist <- tempfile("test-wflow_open-nonexist-")
+  on.exit(unlink(location_nonexist, recursive = TRUE, force = TRUE), add = TRUE)
+  testfile2 <- file.path(location_nonexist, "nonexist.Rmd")
+
+  outfile <- wflow_open(c(testfile1, testfile2), change_wd = TRUE,
+                        open_file = FALSE, project = NULL)
+  expect_true(all(file.exists(outfile)))
+  expect_identical(outfile, workflowr:::relative(c(testfile1, testfile2)))
+})
+
+# Errors -----------------------------------------------------------------------
+
+test_that("wflow_open rejects filenames without Rmd or rmd extension", {
+
+  expect_error(wflow_open(file.path(p$analysis, "invalid-ext.md"), change_wd = FALSE, open_file = FALSE,
+                          project = site_dir),
+               "R Markdown files must have the extension Rmd or rmd.")
+  expect_error(wflow_open("no-ext", change_wd = FALSE, open_file = FALSE,
+                          project = site_dir),
+               "R Markdown files must have the extension Rmd or rmd.")
+  expect_silent(wflow_open(file.path(p$analysis, "valid-ext.Rmd"), change_wd = FALSE, open_file = FALSE,
+                           project = site_dir))
+  expect_silent(wflow_open(file.path(p$analysis, "valid-ext.rmd"), change_wd = FALSE, open_file = FALSE,
+                           project = site_dir))
+})